# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vantor API",
    "version": "0.8.0-one-stop"
  },
  "paths": {
...
    "/api/v1/contracts": {
      "get": {
        "tags": [
          "contracts"
        ],
        "summary": "List Contracts",
        "operationId": "list_contracts_api_v1_contracts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Cursor"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          },
          {
            "name": "expiring",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Expiring"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Contracts Api V1 Contracts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Create Contract",
        "operationId": "create_contract_api_v1_contracts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Contract Api V1 Contracts Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}": {
      "get": {
        "tags": [
          "contracts"
        ],
        "summary": "Get Contract",
        "operationId": "get_contract_api_v1_contracts__contract_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Contract Api V1 Contracts  Contract Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}/status": {
      "patch": {
        "tags": [
          "contracts"
        ],
        "summary": "Move Contract",
        "operationId": "move_contract_api_v1_contracts__contract_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractStatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Move Contract Api V1 Contracts  Contract Id  Status Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}/obligations": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Add Obligation",
        "operationId": "add_obligation_api_v1_contracts__contract_id__obligations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObligationIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Add Obligation Api V1 Contracts  Contract Id  Obligations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/roll-expiry": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Roll Expiry",
        "operationId": "roll_expiry_api_v1_contracts_roll_expiry_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Roll Expiry Api V1 Contracts Roll Expiry Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/contracts/{contract_id}/sign": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Sign Contract",
        "description": "Sign-off record. Internal = authenticated click; esign = external provider\nenvelope (provider + envelope_id required; delivery verified by the adapter).",
        "operationId": "sign_contract_api_v1_contracts__contract_id__sign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Sign Contract Api V1 Contracts  Contract Id  Sign Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}/match": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Run Match",
        "description": "Run the 11-dim deterministic match over server-side rows; store the run.",
        "operationId": "run_match_api_v1_contracts__contract_id__match_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Run Match Api V1 Contracts  Contract Id  Match Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContractIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2,
            "title": "Code"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 2,
            "title": "Title"
          },
          "supplier_id": {
            "type": "string",
            "title": "Supplier Id",
            "default": ""
          },
          "contract_type": {
            "type": "string",
            "title": "Contract Type",
            "default": "supply"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": ""
          },
          "value_minor": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Value Minor",
            "default": 0
          },
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "default": ""
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "default": ""
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "code",
          "title"
        ],
        "title": "ContractIn"
      },
      "ContractStatusIn": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ContractStatusIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchIn": {
        "properties": {
          "po_id": {
            "type": "string",
            "minLength": 1,
            "title": "Po Id"
          },
          "invoice_id": {
            "type": "string",
            "minLength": 1,
            "title": "Invoice Id"
          }
        },
        "type": "object",
        "required": [
          "po_id",
          "invoice_id"
        ],
        "title": "MatchIn"
      },
...
      "ObligationIn": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 2,
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "open"
          },
          "due_date": {
            "type": "string",
            "title": "Due Date",
            "default": ""
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ObligationIn"
      },
...
      "SignIn": {
        "properties": {
          "method": {
            "type": "string",
            "title": "Method",
            "default": "internal"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": ""
          },
          "envelope_id": {
            "type": "string",
            "title": "Envelope Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "SignIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}